requirements.txt
langchain
openai
weaviate-client
apscheduler
selenium
sentence-transformers
undetected-chromedriver 
flask
python-dotenv